+++ app/src/main/java/com/catpuppyapp/puppygit/screen/RemoteListScreen.kt
@@ -48,9 +48,10 @@ import com.catpuppyapp.puppygit.compose.BottomSheetItem
 import com.catpuppyapp.puppygit.compose.ConfirmDialog
 import com.catpuppyapp.puppygit.compose.CopyableDialog
 import com.catpuppyapp.puppygit.compose.CreateRemoteDialog
+import com.catpuppyapp.puppygit.compose.FetchRemotesDialog
 import com.catpuppyapp.puppygit.compose.FilterTextField
+import com.catpuppyapp.puppygit.compose.FullScreenScrollableColumn
 import com.catpuppyapp.puppygit.compose.GoToTopAndGoToBottomFab
-import com.catpuppyapp.puppygit.compose.LoadingTextSimple
 import com.catpuppyapp.puppygit.compose.LongPressAbleIconBtn
 import com.catpuppyapp.puppygit.compose.MyLazyColumn
 import com.catpuppyapp.puppygit.compose.PullToRefreshBox
@@ -297,45 +298,18 @@ fun RemoteListScreen(
     }
 
     if(showFetchAllDialog.value) {
-        ConfirmDialog(
-            title = stringResource(id = R.string.fetch_all),
-            text = stringResource(id = R.string.fetch_all_are_u_sure),
-            onCancel = { showFetchAllDialog.value = false }
-        ) {
-            showFetchAllDialog.value=false
-            doJobThenOffLoading(loadingOn, loadingOff, activityContext.getString(R.string.fetching_all)) {
-                try {
-                    if(list.value.isNotEmpty()) {  //remote列表如果是空就不用fetch all了
-                        //remote名和凭据组合的列表
-                        val remoteCredentialList = Libgit2Helper.genRemoteCredentialPairList(list.value, AppModel.dbContainer.credentialRepository,
-                            requireFetchCredential = true, requirePushCredential = false)
-
-                        Repository.open(curRepo.value.fullSavePath).use { repo ->
-                            //fetch all
-                            Libgit2Helper.fetchRemoteListForRepo(repo, remoteCredentialList, curRepo.value)
-                            //显示成功通知
-                            Msg.requireShow(activityContext.getString(R.string.fetch_all_success))
-                        }
-                    }else {  // remotes列表为空，无需执行操作
-                        Msg.requireShowLongDuration(activityContext.getString(R.string.err_remote_list_is_empty))
-                    }
-
-                }catch (e:Exception){
-                    val errMsg = "fetch all err: "+e.localizedMessage
-                    Msg.requireShowLongDuration(errMsg)
-                    createAndInsertError(curRepo.value.id, errMsg)
-
-                    MyLog.e(TAG, "fetch all err: "+e.stackTraceToString())
-
-                }finally {
-                    changeStateTriggerRefreshPage(needRefresh)
-                }
-
-
-            }
-
-        }
+        FetchRemotesDialog(
+            title = stringResource(R.string.fetch_all),
+            text = stringResource(R.string.fetch_all_are_u_sure),
+            remoteList = list.value,
+            closeDialog = { showFetchAllDialog.value = false },
+            curRepo = curRepo.value,
+            loadingOn = loadingOn,
+            loadingOff = loadingOff,
+            refreshPage = { changeStateTriggerRefreshPage(needRefresh) },
+        )
     }
+
     if(showSetUrlDialog.value) {
         ConfirmDialog(
             okBtnEnabled = isPushUrl.value || urlTextForSetUrlDialog.value.isNotEmpty(),
@@ -870,8 +844,10 @@ fun RemoteListScreen(
             onRefresh = { changeStateTriggerRefreshPage(needRefresh) }
         ) {
 
-            if(isLoading.value) {
-                LoadingTextSimple(contentPadding = contentPadding, text = loadingText.value)
+            if(isLoading.value || list.value.isEmpty()) {
+                FullScreenScrollableColumn(contentPadding) {
+                    Text(if(isLoading.value) loadingText.value else stringResource(R.string.item_list_is_empty))
+                }
             }else {
                 //根据关键字过滤条目
                 val keyword = filterKeyword.value.text.lowercase()  //关键字
@@ -962,13 +938,14 @@ fun RemoteListScreen(
     LaunchedEffect(needRefresh.value) {
         try {
             if(repoId.isNotBlank()) {
-                doJobThenOffLoading {
-                    list.value.clear()
+                doJobThenOffLoading(loadingOn, loadingOff, activityContext.getString(R.string.loading)) {
                     val repoDb = AppModel.dbContainer.repoRepository
                     val repoFromDb = repoDb.getById(repoId)
-                    if(repoFromDb==null) {
+                    if(repoFromDb == null) {
+                        Msg.requireShowLongDuration(activityContext.getString(R.string.repo_id_invalid))
                         return@doJobThenOffLoading
                     }
+
                     curRepo.value = repoFromDb
 
                     val remoteDb = AppModel.dbContainer.remoteRepository
@@ -979,6 +956,7 @@ fun RemoteListScreen(
 //                        updateRemoteDtoList(repo, listFromDb)
 //                    }
 
+                    list.value.clear()
                     list.value.addAll(listFromDb)
 
                     triggerReFilter(filterResultNeedRefresh)