+++ CHANGELOG.md
@@ -4,11 +4,17 @@
 1.1.2.7v103 - 20250606:
 - fix: sometimes commit item not update after reset
 - add Thanks list in About screen
+- support fetch remotes at Branches screen
+- update depend libs (libgit2 to 1.9.1, openssl to 3.5.0)
+- Diff: optimize default match strategy
 
 
 
 - 修复提交条目在重置后不更新的bug
 - 关于页面添加感谢列表
+- 支持在分支页面下载远程仓库
+- 更新依赖库 (libgit2 to 1.9.1, openssl to 3.5.0)
+- Diff: 优化默认匹配策略
 
 
 ---